requirements.txt
asgiref==3.6.0
certifi==2022.12.7
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
dnspython==2.2.1
psycopg2-binary==2.9.5
pymongo==3.12.3
pytz==2022.7
sqlparse==0.2.4
